# docs/api/openapi.yaml
{
  "openapi": "3.1.1",
  "info": {
    "title": "ocfleet read-only observation API",
    "version": "0.3.0",
    "description": "Read-only, low-sensitive observation API over an existing controller SQLite database. It does not call agents, run scheduler jobs, change trust, mutate retention, or expose raw logs/config/cert/session/user/client data."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Default loopback listener for local operator workflows"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {}
  ],
...
  "paths": {
...
    "/api/v1/alerts": {
      "get": {
        "operationId": "listV1Alerts",
        "summary": "Cursor-paginated alerts with reason and half-open time filters",
        "parameters": [
          { "name": "from", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } },
          { "name": "to", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 10000 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string", "maxLength": 2048 } },
          { "name": "state", "in": "query", "schema": { "type": "string", "enum": ["open", "silenced", "resolved"] } },
          { "name": "severity", "in": "query", "schema": { "type": "string", "enum": ["warning", "critical"] } },
          { "name": "node_id", "in": "query", "schema": { "type": "string", "maxLength": 128 } },
          { "name": "reason", "in": "query", "schema": { "type": "string", "maxLength": 128 } }
        ],
        "responses": { "200": { "$ref": "#/components/responses/V1Page" }, "304": { "description": "Representation has not changed" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
    "/api/v1/alerts/{dedupe_key_or_alert_id}": {
      "get": {
        "operationId": "getV1Alert",
        "summary": "Stable v1 single alert projection",
        "parameters": [{ "name": "dedupe_key_or_alert_id", "in": "path", "required": true, "schema": { "type": "string", "maxLength": 256 } }],
        "responses": { "200": { "$ref": "#/components/responses/V1Single" }, "304": { "description": "Representation has not changed" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
    "/jobs": {
      "get": {
        "operationId": "listJobs",
        "summary": "Stored observability jobs",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JobList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "operationId": "getJob",
        "summary": "Single stored observability job",
        "parameters": [
          {
            "$ref": "#/components/parameters/JobId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Job"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/runs": {
      "get": {
        "operationId": "listRuns",
        "summary": "Stored scheduler run summaries",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/JobIdQuery"
          },
          {
            "$ref": "#/components/parameters/RunStatus"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RunList"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/runs/{run_id}": {
      "get": {
        "operationId": "getRun",
        "summary": "Single stored scheduler run summary",
        "parameters": [
          {
            "$ref": "#/components/parameters/RunId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Run"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/observations": {
      "get": {
        "operationId": "listObservations",
        "summary": "Low-sensitive stored observation summaries",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/NodeIdQuery"
          },
          {
            "$ref": "#/components/parameters/MethodQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ObservationList"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/observations/{observation_id}": {
      "get": {
        "operationId": "getObservation",
        "summary": "Single low-sensitive stored observation summary",
        "parameters": [
          {
            "$ref": "#/components/parameters/ObservationId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Observation"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/alerts": {
      "get": {
        "operationId": "listAlerts",
        "summary": "Stored alert event summaries",
        "parameters": [
          {
            "$ref": "#/components/parameters/AlertState"
          },
          {
            "$ref": "#/components/parameters/AlertSeverity"
          },
          {
            "$ref": "#/components/parameters/NodeIdQuery"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AlertList"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/alerts/{dedupe_key_or_alert_id}": {
      "get": {
        "operationId": "getAlert",
        "summary": "Single stored alert event summary",
        "parameters": [
          {
            "$ref": "#/components/parameters/AlertLookup"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Alert"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/audit/export": {
      "get": {
        "operationId": "getAuditExport",
        "summary": "Bounded low-sensitive audit export rows",
        "parameters": [
          {
            "$ref": "#/components/parameters/From"
          },
          {
            "$ref": "#/components/parameters/To"
          },
          {
            "$ref": "#/components/parameters/Redact"
          },
          {
            "$ref": "#/components/parameters/MaxRows"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuditList"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Required for non-loopback listeners via a private --auth-token-file and resolves only to the viewer role. The API is plain HTTP: non-loopback deployments must use a TLS-terminating reverse proxy plus firewall/private-network controls, and must not expose port 8080 directly to the public Internet. Loopback local-operator listeners may run unauthenticated."
      }
    },
    "parameters": {
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 10000,
          "default": 50
        }
      },
...
      "NodeIdQuery": {
        "name": "node_id",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/SafeIdentifier"
        }
      },
      "JobId": {
        "name": "job_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/SafeIdentifier"
        }
      },
      "JobIdQuery": {
        "name": "job_id",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/SafeIdentifier"
        }
      },
      "RunId": {
        "name": "run_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/SafeIdentifier"
        }
      },
      "ObservationId": {
        "name": "observation_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/SafeIdentifier"
        }
      },
      "AlertLookup": {
        "name": "dedupe_key_or_alert_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/SafeIdentifier"
        }
      },
      "MethodQuery": {
        "name": "method",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/SafeIdentifier"
        }
      },
      "RunStatus": {
        "name": "status",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "running",
            "succeeded",
            "failed",
            "skipped"
          ]
        }
      },
      "AlertState": {
        "name": "state",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "open",
            "silenced",
            "resolved"
          ]
        }
      },
      "AlertSeverity": {
        "name": "severity",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "warning",
            "critical"
          ]
        }
      },
      "From": {
        "name": "from",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "To": {
        "name": "to",
        "in": "query",
        "required": true,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "Redact": {
        "name": "redact",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "none",
            "default",
            "strict"
          ],
          "default": "default"
        }
      }
    },
    "responses": {
      "V1Page": {
        "description": "Signed cursor page with conditional caching",
        "headers": { "ETag": { "schema": { "type": "string" } }, "X-Request-ID": { "schema": { "type": "string", "maxLength": 128 } } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/V1GenericPageEnvelope" } } }
      },
      "V1Single": {
        "description": "Stable v1 single-record envelope",
        "headers": { "ETag": { "schema": { "type": "string" } }, "X-Request-ID": { "schema": { "type": "string", "maxLength": 128 } } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/V1SingleEnvelope" } } }
      },
...
      "JobList": {
        "description": "Bounded observability job list",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/JobListResponse"
            }
          }
        }
      },
      "Job": {
        "description": "Single observability job projection",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/JobResponse"
            }
          }
        }
      },
      "RunList": {
        "description": "Bounded scheduler run list",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RunListResponse"
            }
          }
        }
      },
      "Run": {
        "description": "Single scheduler run projection",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RunResponse"
            }
          }
        }
      },
      "ObservationList": {
        "description": "Bounded low-sensitive observation list",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ObservationListResponse"
            }
          }
        }
      },
      "Observation": {
        "description": "Single low-sensitive observation projection",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ObservationResponse"
            }
          }
        }
      },
      "AlertList": {
        "description": "Bounded low-sensitive alert list",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AlertListResponse"
            }
          }
        }
      },
      "Alert": {
        "description": "Single low-sensitive alert projection",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AlertResponse"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Invalid bounded query or identifier",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Item not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
...
  }
}